refactor: move contains_point from shape to _shape
